Add Git commit and build info to defines.h. Change window resize to viewport set. Add more modes for computing view offset distance. Line thickness for annotations. Clean up landmark and annotaions windows. Jump to end of landmarks list on adding new one.

+++ src/AntropyApp.h
@@ -49,7 +49,7 @@ public:
     void run();
 
     /// Resize the framebuffer
-    void resize( int width, int height );
+    void resize( float width, float height );
 
     /// Render one frame
     void render();
